Move Kraken reading logic to separate class
