+++ swift/Sources/main.swift
@@ -6,120 +6,14 @@ import JSONDecoding
 import KrakenAPI
 import SwiftCSV
 
-private let csv: CSV = try CSV<Named>(url: URL(fileURLWithPath: "./ledgers.csv"))
+private let krakenReader = KrakenCSVReader()
+private let krakenTransactions = try await krakenReader.read(filePath: "../data/Kraken.csv")
 
-private enum AssetType {
-    case fiat
-    case crypto
+for transaction in krakenTransactions.filter({ $0.type == .Withdrawal }) {
+    print("\(transaction.provider) \(transaction.date) \(transaction.amount) \(transaction.asset.name)")
 }
 
-private struct Asset {
-    let name: String
-    let type: AssetType
-
-    init(fromTicker ticker: String) {
-        switch ticker {
-        case "XXBT":
-            self.name = "BTC"
-            self.type = .crypto
-        case "XXDG":
-            self.name = "DOGE"
-            self.type = .crypto
-        case let a where a.starts(with: "X"):
-            self.name = String(a.dropFirst())
-            self.type = .crypto
-        case let a where a.starts(with: "Z"):
-            self.name = String(a.dropFirst())
-            self.type = .fiat
-        default:
-            self.name = ticker
-            self.type = .crypto
-        }
-    }
-}
-
-private struct LedgerEntry {
-    let txId: String
-    let refId: String
-    let time: String
-    let type: String
-    let asset: Asset
-    let amount: Decimal
-}
-
-private struct Trade {
-    let from: Asset
-    let fromAmount: Decimal
-    let to: Asset
-    let toAmount: Decimal
-    let rate: Decimal
-
-    init?(fromLedgers entries: [LedgerEntry]) {
-        if entries.count < 2 {
-            return nil
-        }
-
-        if entries[0].amount < 0 {
-            self.from = entries[0].asset
-            self.fromAmount = -entries[0].amount
-            self.to = entries[1].asset
-            self.toAmount = entries[1].amount
-        } else {
-            self.from = entries[1].asset
-            self.fromAmount = -entries[1].amount
-            self.to = entries[0].asset
-            self.toAmount = entries[0].amount
-        }
-
-        self.rate = self.fromAmount / self.toAmount
-    }
-}
-
-private let rateFormatterFiat = NumberFormatter()
-rateFormatterFiat.maximumFractionDigits = 4
-rateFormatterFiat.minimumFractionDigits = 0
-
-private let btcFormatter = NumberFormatter()
-btcFormatter.maximumFractionDigits = 8
-btcFormatter.minimumFractionDigits = 8
-
-private let rateFormatterCrypto = NumberFormatter()
-rateFormatterCrypto.maximumFractionDigits = 10
-rateFormatterCrypto.minimumFractionDigits = 0
-
-private func printTrade(entries: [LedgerEntry]) {
-    guard let trade = Trade(fromLedgers: entries) else {
-        return
-    }
-
-    let rateFormatter = trade.from.type == .fiat ? rateFormatterFiat : rateFormatterCrypto
-
-    if trade.from.name != "EUR" || trade.to.name != "BTC" {
-        return
-    }
-    print("Traded", trade.fromAmount, trade.from.name, "for", trade.toAmount, trade.to.name, "@", rateFormatter.string(for: trade.rate)!)
-}
-
-private var ledgers = [LedgerEntry]()
-private var ledgersByRefId = [String: [LedgerEntry]]()
-// "txid","refid","time","type","subtype","aclass","asset","amount","fee","balance"
-try csv.enumerateAsDict { dict in
-    let entry = LedgerEntry(txId: dict["txid"] ?? "",
-                            refId: dict["refid"] ?? "",
-                            time: dict["time"] ?? "",
-                            type: dict["type"] ?? "",
-                            asset: Asset(fromTicker: dict["asset"] ?? ""),
-                            amount: Decimal(string: dict["amount"] ?? "0") ?? 0)
-    ledgers.append(entry)
-    if !entry.refId.isEmpty {
-        ledgersByRefId[entry.refId, default: []].append(entry)
-    }
-}
-
-private let ledgersGroupedByRefId = ledgersByRefId.values.filter { $0.count > 1 }
-
-print(ledgers.count)
-print(ledgersByRefId.count)
+exit(0)
 
 // Addresses that are plausibly part of the wallet of the user or have been in the past
 private var internalAddresses = Set<Address>(knownAddresses)
@@ -164,7 +58,7 @@ private func isManualTransaction(_ transaction: ElectrumTransaction) -> Bool {
     return transaction.vout.count <= 2
 }
 
-private func hasExternalAddressesInputs(_ transaction: Transaction) -> Bool {
+private func hasExternalAddressesInputs(_ transaction: OnchainTransaction) -> Bool {
     return transaction.vin.contains(where: { externalAddresses.contains($0.address) })
 }
 
@@ -231,7 +125,7 @@ private func round(no: Int) async -> Int {
         .compactMap { $0 }
     _ = await retrieveAndStoreTransactions(txIds: refTransactionIds)
 
-    var transactions = [Transaction]()
+    var transactions = [OnchainTransaction]()
     for rawTransaction in rootTransactions {
         transactions.append(await buildTransaction(
             transaction: rawTransaction,
@@ -257,9 +151,9 @@ private func round(no: Int) async -> Int {
                 print(vin.address.id)
             }
         } else if case .deposit(let amount) = transaction.type {
-            if hasExternalAddressesInputs(transaction) {
-                print("🎉 Marked external from previous round!!")
-            }
+//            if hasExternalAddressesInputs(transaction) {
+//                print("🎉 Interacting with external address!")
+//            }
             print("External Service Deposit", satsToBtc(amount), transaction.txid)
         }
     }
@@ -267,6 +161,8 @@ private func round(no: Int) async -> Int {
     // Returns the number of addresses added to the internal addresses.
     return internalAddresses.count - internalAddressesList.count
 }
+
+private let btcFormatter = createNumberFormatter(minimumFractionDigits: 8, maximumFranctionDigits: 8)
 
 func satsToBtc(_ amount: Int) -> String {
     btcFormatter.string(from: Double(amount) / 100000000 as NSNumber)!
@@ -277,9 +173,9 @@ func btcToSats(_ amount: Double) -> Int {
 }
 
 @MainActor
-func buildTransaction(transaction: ElectrumTransaction, address: Address) async -> Transaction {
+func buildTransaction(transaction: ElectrumTransaction, address: Address) async -> OnchainTransaction {
     var totalIn = 0
-    var transactionVin = [TransactionVin]()
+    var transactionVin = [OnchainTransaction.Vin]()
     for vin in transaction.vin {
         guard let vinTxId = vin.txid else {
             continue
@@ -304,7 +200,7 @@ func buildTransaction(transaction: ElectrumTransaction, address: Address) async 
         }
 
         totalIn += sats
-        transactionVin.append(TransactionVin(
+        transactionVin.append(OnchainTransaction.Vin(
             txid: vinTxId,
             voutIndex: voutIndex,
             sats: sats,
@@ -313,7 +209,7 @@ func buildTransaction(transaction: ElectrumTransaction, address: Address) async 
     }
 
     var totalOut = 0
-    var transactionVout = [TransactionVout]()
+    var transactionVout = [OnchainTransaction.Vout]()
     for vout in transaction.vout {
         guard let voutAddress = vout.scriptPubKey.address else {
             continue
@@ -325,7 +221,7 @@ func buildTransaction(transaction: ElectrumTransaction, address: Address) async 
 
         let sats = btcToSats(vout.value)
         totalOut += sats
-        transactionVout.append(TransactionVout(
+        transactionVout.append(OnchainTransaction.Vout(
             sats: sats,
             address: Address(id: voutAddress, scriptHash: voutScriptHash)
         ))
@@ -334,7 +230,7 @@ func buildTransaction(transaction: ElectrumTransaction, address: Address) async 
     let (knownVin, _) = transactionVin.partition { internalAddresses.contains($0.address) }
     let (knownVout, unknownVout) = transactionVout.partition { internalAddresses.contains($0.address) }
 
-    let type: TransactionType = if
+    let type: OnchainTransaction.TransactionType = if
         knownVin.count == transaction.vin.count,
         knownVout.count == transaction.vout.count
     {
@@ -352,7 +248,7 @@ func buildTransaction(transaction: ElectrumTransaction, address: Address) async 
 
     // print("total amount \(Double(totalIn) / 100000000), fee \(totalIn - totalOut) sats")
 
-    return Transaction(
+    return OnchainTransaction(
         txid: transaction.txid,
         time: transaction.time ?? 0,
         rawTransaction: transaction,
@@ -366,7 +262,7 @@ func buildTransaction(transaction: ElectrumTransaction, address: Address) async 
 }
 
 var no = 1
-while no <= 5 {
+while no <= 10 {
     let addressesAdded = await round(no: no)
     print("------------------------------")
     print("Round \(no) # addresses added: \(addressesAdded)")
@@ -384,45 +280,12 @@ for address in internalAddresses where address.path.count > 0 {
     print()
 }
 
-// let internalAddressesList = internalAddresses.map { $0 }.sorted(by: { $0.id < $1.id })
-// let otherReqs = internalAddressesList.map { JSONRPCRequest.getScripthashHistory(scriptHash: $0.scriptHash) }
-// if let otherRes: [Result<GetScriptHashHistoryResult, JSONRPCError>] = await client.send(requests: otherReqs) {
-//    print("--- OTHER ADDRESSES ---")
-//    for (index, res) in otherRes.enumerated() {
-//        let address = internalAddressesList[index]
-//        print(address.id)
-//
-//        switch res {
-//        case .success(let history):
-//            print(history.map { $0.tx_hash })
-//        case .failure(let error):
-//            if error.historyTooLarge {
-//                print("Moving \(address.id) to external addresses")
-//                print(internalAddresses.count)
-//                internalAddresses.remove(address)
-//                print(internalAddresses.count)
-//                externalAddresses.insert(address)
-//            } else {
-//                print("🚨 history request failed", error)
-//            }
-//        }
-//    }
-// } else {
-//    print("🚨 Failed to fetch history of other addresses")
-// }
-
 /*
  BRAIN DUMP:
- - implement error handling in JSONRPCClient to handle this shape of response:
-    `{"jsonrpc":"2.0","error":{"code":1,"message":"history too large"},"id":123}`
- - when error message is "history too large" (code: 1) it means the address has too many transactions, so the
-   original transaction that was marked as "Manual" is not manual at all, it's part of an external service
- - the thing that was considered a manual deposit must become an external service deposit,
-   marked for look up in external data sources (CSV and whatnot)
- - so: tag transactions as manual/external by default using our euristics, but allow it to change and store the change
- - link "other address" with originating transaction to allow updating of transaction tag
- - once cleanup and transaction tagging is complete download all missing transactions by txid and recompute
-   external vs manual transactions, rinse and repeat
- - open question: where to store "other addresses"? with knownAddresses? as its separate thing?
-   we need a more general solution that can scale beyond knownAddresses / other addresses to multiple wallets maybe?
+ - the machinery above works!! but, it needs deposit amounts and dates so it can match them with
+   what is found on chain and tag external addresses. I've done this manually for a single address
+   (albeith a pretty complex one, the oldest probably) with great success.
+ - so next step is ingest all CSV and make them available to be matched
+ - what to do with Coinbase's pdf? copy-pasta or something.. I don't know...
+ - ‼️ GOAL: have this work without the hardcoded Celsius and Coinbase addresses, but with that as part of the output
  */
